chore: general file refactor
