+++ .tool-versions
@@ -1,1 +1,1 @@
-nodejs latest
+nodejs 23.3.0
