+++ src/utils/extract-credits.ts
